[SES-47] Isolating models from views
