+++ src/core/models/custom-chart-item.model.ts
@@ -1,0 +1,4 @@
+export interface CustomChartItemModel {
+  label?: string,
+  value: number,
+}
