+++ src/core/models/facilitator.model.ts
@@ -1,0 +1,5 @@
+export interface FacilitatorModel {
+  name: string,
+  imageUrl?: string,
+  id?: string,
+}
